+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+- Add support for new `invoke` calls from smart contracts in protocol version 7:
+  - query the contract module reference for a given contract address
+  - query the contract name for a given contract address
+
 ## 6.3.0
 
 - Fix a bug where `GetBlockPendingUpdates` fails to report pending updates to the finalization